support OutputInteface from symfony 5 and symfony >6
